demo de sistema solar

+++ engine/Vec3.cpp
@@ -57,7 +57,7 @@ float Vec3::length() {
     return sqrtf(powf(m_x, 2) + powf(m_y, 2) + powf(m_z, 2));
 }
 
-void Vec3::set(int x, int y, int z){
+void Vec3::set(float x, float y, float z){
     m_x = x;
     m_y = y;
     m_z = z;
