+++ engine/includes/Vec3.h
@@ -19,7 +19,7 @@ public:
     float getX();
     float getY();
     float getZ();
-    void set(int x, int y, int z);
+    void set(float x, float y, float z);
     Vec3 normalize();
     float length();
 };
